Began adding support for multiple gravitational bodies
Using static meshes tagged "Gravity" to find the masses that should be
counted. It works except for an apparent error with signs because gravity
works opposite the way it should for the 2nd body. Jumping under the 2nd
body, the player hits the 1st body hard. On the opposite side of the 1st
body, jumping can put the player into temporary orbit.

Accidentally retextured the universe to show the Milky Way.