Gravity Manager for all added
Isolated some of the gravity code into GravityManager.cpp. It seems to
work well for objects other than the player. The player seems intially
attracted to the center of mass, but then things go backward. I haven't
been able to tell if it is a sign problem or if for is being applied to
wrong reference frame (local, world, relative). The code is all rather
ugly and non-optimised at this point so this is mostly a safty commit.